requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
requests
pytest
flask_cors